% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_top_words.R
\name{append_top_words}
\alias{append_top_words}
\title{Top Words for PubMed Abstracts}
\usage{
append_top_words(abstracts_dataframe, num_words = 30)
}
\arguments{
\item{abstracts_dataframe}{A DataFrame returned from the \code{get_pubmed_abstracts} function.
Must already have tf-idf scores calculated from \code{append_tfidf}}
}
\value{
A DataFrame.
}
\description{
\code{append_tfidf} Finds words with the top tf-idf scores for each abstract entry in the dataframe.
}
\details{
\code{append_tfidf} Given a dataframe returned from the \code{get_pubmed_abstracts} function that
already has tf-idf scores calculated from \code{append_tfidf}, append a column to the DataFrame
of the words with the top tf-idf scores.
}
\examples{
abstracts_with_topwordsf <- append_top_words(abstracts_df_with_tfidf)

}
\seealso{
\code{get_pubmed_abstracts} \code{\link{append_tfidf}}
}
