% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_total_information.R
\name{append_total_information}
\alias{append_total_information}
\title{Calculate Total Tf-Idf Scores for All Documents}
\usage{
append_total_information(abstracts_dataframe)
}
\arguments{
\item{abstracts_dataframe}{A DataFrame returned from the \code{get_pubmed_abstracts} function.  Must already have tf-idf scores calculated from \code{append_tfidf}}
}
\value{
A DataFrame.
}
\description{
\code{append_total_information} Adds up total tf-idf scores among all abstracts.
}
\details{
\code{append_total_information} Given a dataframe returned from the \code{get_pubmed_abstracts} function that already has tf-idf scores calculated from \code{append_tfidf}, append a column to the DataFrame of the total tf-idf scores.
}
\examples{
abstracts_df_with_total <- append_top_words(abstracts_df_with_tfidf)

}
\seealso{
\code{get_pubmed_abstracts} \code{\link{append_tfidf}}
}
