% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arabidopsis_data.R
\docType{data}
\name{arabidopsis_data}
\alias{arabidopsis_data}
\title{PubMed Abstracts corresponding to a search for terms "arabidopsis thaliana gibberellin"}
\format{An object of class \code{"data.frame"};}
\usage{
data(arabidopsis_data)
}
\description{
Data from function \code{get_pubmed_abstracts} with arguments "arabidopsis thaliana gibberellin", default number of results returned
}
\examples{
arabidopsis.data$abstract
names(arabidopsis.data)
}
\keyword{datasets}
