% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_similar.R
\name{get_most_similar}
\alias{get_most_similar}
\title{Most Similar Documents}
\usage{
get_most_similar(tfidf_vector, abstracts_dataframe)
}
\arguments{
\item{abstracts_dataframe}{A DataFrame returned from the \code{get_pubmed_abstracts} function.  Must already have tf-idf scores calculated from \code{append_tfidf}.}

\item{tf-idf_vector}{A numeric vector of tf-idf weights that has the same number of columns as those in the abstracts DataFrame.}
}
\value{
A DataFrame.
}
\description{
\code{get_most_similar} Finds the documents with the highest cosine-similarity score among their tf-idf vectors.
}
\details{
\code{append_tfidf} Given a dataframe returned from the \code{get_pubmed_abstracts} function that already has tf-idf scores calculated from \code{append_tfidf}, and a vector of tf-idf scores that have the same number of columns, return the indices of the documents that have the highest cosine-similarity scores in descending order.
}
\examples{
indices <-get_most_similar(abstracts_df_with_tfidf,tfidf_vector)

}
\seealso{
\code{get_pubmed_abstracts} \code{\link{append_tfidf}}
}
