% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pubmed_abstracts.R
\name{get_pubmed_abstracts}
\alias{get_pubmed_abstracts}
\title{Abstracts and Document Information from PubMed}
\usage{
get_pubmed_abstracts(query_string, retmax = 20)
}
\arguments{
\item{query_string}{A string to be searched on PubMed}

\item{retmax}{Maximum number of results returned}
}
\value{
A DataFrame.
}
\description{
\code{get_pubmed_abstracts} Returns information on abstracts of documents retrieved from PubMed
}
\details{
\code{get_pubmed_abstracts} Takes a query string and returns a dataframe containing the text of the abstract and other information on documents returned from PubMed.
}
\examples{
query_string <- gene expression cancer"
abstracts_df <- get_pubmed_abstracts(query_string)
abstracts_df$abstract[1] <- "Membrane type-1 matrix metalloproteinase (MT1-MMP) possesses....
}
\seealso{
\code{\link{article_to_df}}
}
