% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_word_score.R
\name{plot_word_score}
\alias{plot_word_score}
\title{Plot Bar Graph of Tf-Idf Scores for an Abstract}
\usage{
plot_word_score(tfidf_vector, topn = 25)
}
\arguments{
\item{topn}{Number of top words to plot, default 25.}

\item{tf-idf_vector}{A numeric vector of tf-idf weights that has the same number of columns as those in the abstracts DataFrame.}

\item{title}{Title of the document that hte tf-idf vector belongs to.}
}
\description{
\code{plot_word_score} Produces a bar graph of the words with the highest tf-idf scores.
}
\details{
\code{plot_word_score} Given a vector of tf-idf weights, and the title of the document the vector belongs to, produces a bar plot of the words with the highest tf-idf weights.
}
\examples{
plot_word_score(tfidf_vector,"Review of Current Machine Learning Approaches in Plant Biology",35)

}
\seealso{
\code{get_pubmed_abstracts} \code{\link{append_tfidf}}
}
