% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_and_plot.R
\name{reduce_and_plot}
\alias{reduce_and_plot}
\title{Scatter Plot of Tf-Idf}
\usage{
reduce_and_plot(abstracts_dataframe)
}
\arguments{
\item{abstracts_dataframe}{DataFrame produced from \code{get_pubmed_abstracts}.  Must have tf-idf vectors already from \code{append_tfidf}.}
}
\description{
\code{plot_word_score} Produces a scatter plot for all documents with Tf-Idf vectors.
}
\details{
\code{plot_word_score}Given a dataframe from the \code{get_pubmed_abstracts} function with tf-idf weights from the \code{append_tfidf} function, produce a scatterplot using a PCA reduced coordinate of all abstracts present in the DataFrame.
}
\examples{
reduce_and_plot(abstracts_df_with_tfidf)

}
\seealso{
\code{get_pubmed_abstracts} \code{\link{append_tfidf}}
}
